Simplified driver interface by moving code to other modules
