+++ nvflare/fuel/f3/drivers/uds_driver.py
@@ -33,7 +33,7 @@ from typing import List, Dict, Any
 
 from nvflare.fuel.f3.comm_error import CommError
 from nvflare.fuel.f3.drivers.base_driver import BaseDriver
-from nvflare.fuel.f3.drivers.driver import Driver, DriverParams, Connector, DriverCap
+from nvflare.fuel.f3.drivers.driver import Driver, Connector
 from nvflare.fuel.f3.drivers.socket_conn import ConnectionHandler, SocketConnection
 
 log = logging.getLogger(__name__)
@@ -68,7 +68,7 @@ class UdsDriver(BaseDriver):
 
     @staticmethod
     def supported_transports() -> List[str]:
-        return ["uds"]
+        return ["ouds"]
 
     @staticmethod
     def capabilities() -> Dict[str, Any]:
